Move to pytest
Use pytest instead of nose to run tests.
