+++ .travis.yml
@@ -8,7 +8,7 @@ python:
 
 # command to run tests
 script:
-    - "python setup.py nosetests"
+    - "python setup.py pytest"
 
 
 # EOF
