--- apiramid/test/__init__.py
@@ -1,5 +1,0 @@
-""" Tests module
-"""
-
-
-# EOF
