+++ tests/test_content_types.py
@@ -4,12 +4,12 @@
 
 import json
 import os
+
+import apiramid
 import pyramid.testing
 import unittest
 import webtest
 import xmltodict
-
-from . import content_types
 
 
 PACKAGE_NAME = 'apiramid'
@@ -21,12 +21,70 @@ DOCUMENT_PATH = os.path.join(
 )
 
 
+
+@apiramid.endpoint.EndpointDefaults(path='/capitalize', deserialize=True)
+class Capitalize(object):
+
+    def __init__(self, context, request):
+        self.request = request
+        return None
+
+    def capitalize(self):
+        text = self.request.dict_body.get('text', '')
+        result = {
+            'text': text.capitalize(),
+        }
+        return result
+
+    @apiramid.endpoint.Endpoint(request_method='PUT')
+    def put(self):
+        result = self.capitalize()
+        return pyramid.httpexceptions.HTTPOk(
+            body=result,
+        )
+
+
+class XmlRendererFactory:
+
+    MEDIA_TYPE = 'application/xml'
+
+    def __init__(self, info):
+        pass
+
+    def __call__(self, value, system):
+        request = system['request']
+        response = request.response
+        response.content_type = self.MEDIA_TYPE
+        result = xmltodict.unparse(value)
+        return result
+
+
+class XmlDeserializerFactory:
+
+    def __init__(self, info):
+        pass
+
+    def __call__(self, value, system):
+        result = xmltodict.parse(value)
+        return result
+
+
+class JsonDeserializerFactory:
+
+    def __init__(self, info):
+        pass
+
+    def __call__(self, value, system):
+        result = system['request'].json_body
+        return result
+
+
 class TestAccept(unittest.TestCase):
 
     def setUp(self):
         self.path = '/v0.0/capitalize'
         self.expected_status = 200
-        self.input_text = 'test text'
+        self.input_text = "test text"
         self.input_dict = {'text': self.input_text}
         self.output_text = self.input_text.capitalize()
         settings = {
@@ -35,16 +93,16 @@ class TestAccept(unittest.TestCase):
         self.config = pyramid.testing.setUp(settings=settings)
         self.config.include(PACKAGE_NAME)
 
-        self.config.add_renderer('xml', content_types.XmlRendererFactory)
+        self.config.add_renderer('xml', XmlRendererFactory)
         self.config.set_media_renderer('application/xml', 'xml')
 
-        self.config.add_deserializer('xml', content_types.XmlDeserializerFactory)
+        self.config.add_deserializer('xml', XmlDeserializerFactory)
         self.config.set_media_deserializer('application/xml', 'xml')
 
-        self.config.add_deserializer('json', content_types.JsonDeserializerFactory)
+        self.config.add_deserializer('json', JsonDeserializerFactory)
         self.config.set_media_deserializer('application/json', 'json')
 
-        self.config.scan('.content_types')
+        self.config.scan('.')
         self.test_application = webtest.TestApp(self.config.make_wsgi_app())
         return None
